Refactor default values
Reworks the default variables for various roles to make them more
portable. In the future, the roles and playbooks plus group/host vars
will be separate, with the idea that roles will be portable and public,
while the rest of the codebase (including Ansible Vault-protected
secrets) will be in a separate (private) git repository.