Create letsencrypt-dns role
In order to generate SSL certs for non-external (private) servers, the
LetsEncrypt DNS challenge must be used. This is substantively different
than the automatic certbot challenges, and requires temporarily
modifying DNS records. This role issues an LE DNS challenge and then
adds and deletes a Route 53 record to accomplish this.

In the future, consider merging this role into the standard letsencrypt
role, perhaps as a separate tasks file. As it is, there is a fair deal
of overlap, especially when it comes to webserver tasks.